Update README and fix project struct
The `project` struct was the only one that had all pointers, so I reverted that to make things consistent again.

The same goes for the `time.Time` pointers. In some cases the API does not return a time and if the field is not a pointer to `time.Time`, this will generate an error. So to be consistent now all `time.Time` fields are pointer fields.