Implement Epic and Commit discussions
Also implement the correct endpoints for adding, updating and deleting
discussion notes as per the API docs.

Note that the docs at the time of this commit specify an incorrect path
prefix for the commit related endpoints; it should be
/projects/:id/repository/commits/.

Also note that the docs state the commit iids as int, but they are
really string. When a discussion note is returned, the noteable_iid
and noteable_id fields are null.

Finally, all the endpoints for adding a note to a discussion are
documented as taking a note ID, which does not make sense, seeing as how
the caller is adding a new note and is in fact expecting to get such an
ID in the response.

The epics endpoints are GitLab EE only, as indicated by the API
documentation links.

A follow-up commit will correct line lengths for comments and LoCs.

Signed-off-by: Sune Keller <absukl@almbrand.dk>