ListTodo allows for ListOptions
The GitLab API Docs are missing these options, but they are supported as
the GitLab specs show[1], and can be verified for a signed in user[2].

Including them here breaks the interface when users used assignments
without Field names.

To test the change, an PerPage list option was set in the test. This
doesn't influence the test, but does test the interface.

[1]: https://gitlab.com/gitlab-org/gitlab-ce/blob/4325986f25b7523acac49cfaefb7149208ae8553/spec/requests/api/todos_spec.rb#L39
[2]: https://gitlab.com/api/v4/todos?per_page=1