Add StartBranch field into CreateFileOptions and UpdateFileOptions
This field is mandatory if you want to create or
update file onto a new (non-existing) branch.

If the option is missing, then the GitLab API will
return HTTP 400: You can only create or edit files
when you are on a branch.

Note for CreateFile: the GitLab API will not send
proper error message, just silently fail.

Signed-off-by: Alex Szakaly <alex.szakaly@gmail.com>