Return a *bytes.Reader instead of an io.Reader
Change GetJobArtifacts, DownloadArtifactsFile,
DownloadSingleArtifactsFile, and GetTraceFile to return
a *bytes.Reader instead of an io.Reader.

This change makes it simple to e.g. unzip archive files without
needing to first write the data to a file.

I don't see any applicable tests to touch up, I've tested it in a
personal application that uses GetJobArtifacts to download and unpack
artifact archive files and it works as expected.