Add the `Draft` parameter for merge request structs
Currently, in GitLab 14.4.0 the `draft` url parameter is not implemented -
and I'm still unsure whether it should be? You still filter for non-draft
merge requests with `wip=yes`. If GitLab ever removes the `wip` url param
and replaces it with `draft` the typing of this field becomes a difficult
question I think, might be wrong here.

This should be backwards compatible due to the `omitempty` being in both the
`Draft` and `WIP` fields of the merge request structs.

Also adjusted the tests to use `Draft:` instead of `WIP:`. It might be worth
considering having tests for both but that would raise questions on which
behaviour to expect - the removed `WIP:` not working or until it is still in
supported versions to maintain backwards compatibility.