Derive TodoTarget.IDString
Create derived field IDString to work around
https://github.com/xanzy/go-gitlab/issues/1743.

Sometimes GitLab will return a string id instead of an int (e.g. when
target is a commit). This change intercepts a string id before it
causes unmarshaling to fail and redirects it to the new IDString
field. Conversely when marshaling it routes either ID or IDString into
the expected JSON "id" field.